Move ImageListState declaration where it is used (only one place)
